requirements.txt
actionlib==1.12.1
angles==1.9.12
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
beautifulsoup4==4.6.0
blinker==1.4
bondpy==1.8.5
Brlapi==0.6.6
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.5.2
catkin-pkg-modules==0.5.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.4.6
contourpy==1.0.7
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cryptography==2.1.4
cupshelpers==1.0
cv_bridge==1.13.1
cycler==0.10.0
Cython==0.29.33
decorator==4.1.2
default_cfg_fkie==0.8.12
defer==1.0.6
defusedxml==0.7.1
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
distro==1.8.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.19
dynamic_reconfigure==1.6.5
empy==3.3.2
execnet==1.9.0
feedparser==5.2.1
filelock==3.9.0
fonttools==4.38.0
future==0.18.2
futures==3.0.5
gast==0.2.2
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gdown==4.6.0
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
geographiclib==2.0
geopy==2.3.0
gps_common==0.3.4
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
image_geometry==1.13.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
interactive-markers==1.11.5
jetson-stats==4.1.0
Jetson.GPIO==2.0.17
joblib==1.2.0
joint_state_publisher==1.12.15
kdl-parser-py==1.13.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.4.4
language-selector==0.1
laser_geometry==1.6.7
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
llvmlite==0.39.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
master_discovery_fkie==0.8.12
master_sync_fkie==0.8.12
matplotlib==3.7.0
message_filters==1.14.13
mock==3.0.5
node_manager_fkie==0.8.12
numba==0.56.4
numpy==1.23.5
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
onboard==1.4.1
packaging==23.0
PAM==0.4.2
pandas==1.5.3
pbr==5.11.1
Pillow==9.4.0
protobuf==3.0.0
pybind11==2.11.1
pycairo==1.23.0
pycrypto==2.6.1
pycups==1.9.73
PyExifTool==0.5.5
pyflakes==1.6.0
Pygments==2.2.0
PyGObject==3.42.2
PyICU==1.9.8
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.0.9
pyRFC3339==1.0
pyserial==3.5
PySocks==1.7.1
python-apt==1.6.6
python-dateutil==2.8.2
python-debian==0.1.32
python_qt_binding==0.4.4
pytz==2022.7.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.18.4
requests-unixsocket==0.1.5
resource_retriever==1.12.7
roman==2.0.0
rosapi==0.11.16
rosbag==1.14.13
rosboost-cfg==1.14.9
rosbridge-library==0.11.16
rosbridge-server==0.11.16
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospkg==1.4.0
rospkg-modules==1.5.0
rospy==1.14.13
rosserial_arduino==0.8.0
rosserial_client==0.8.0
rosserial_python==0.8.0
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rviz==1.13.30
scikit-learn==1.2.0
scipy==1.10.0
SecretStorage==2.3.1
sensor-msgs==1.12.8
simple-pid==1.0.1
simplejson==3.13.2
six==1.11.0
sklearn==0.0.post1
smach==2.0.3
smach_ros==2.0.3
smbus==1.1.post2
smclib==1.8.5
ssh-import-id==5.7
system-service==0.3
systemd-python==234
termcolor==2.2.0
testresources==2.0.1
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
threadpoolctl==3.1.0
topic_tools==1.14.13
tqdm==4.64.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urdfdom-py==0.4.6
urllib3==1.22
urwid==2.0.1
wadllib==1.3.2
webencodings==0.5
xacro==1.13.20
xkit==0.0.0
youtube_dl==2018.3.14
zipp==3.11.0
zope.interface==4.3.2
